C preprocessor work & a lot more
- Made cast initate the top opearnd
- Made type checking in "intermediate.c" output an error
- Added else intermediate
- Tried fixing constant char preprocessing but constant chars and strings need
to be be done during tokenization
- Fixed typedefs
- Added a header file for compilers that lack default definitions
- Added the first pass optimization step
- Started the C preprocessor

+++ include/frontend/common/parser.h
@@ -20,7 +20,7 @@ type parse_type(char** string_ptr);
 /*
  * This returns the ASCII number of a string.
  */
-i128 get_ascii_number(char* num_string);
+i64 get_ascii_number(char* num_string);
 
 /*
  * This goes through a string and returns true if it is an ASCII number.
